+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* NP-290 Success message
 * NP-214 Link benefits to the fully qualified URL of the public website.
 * NP-301 Textarea control